Make client-side span tagging customisable
I need to be able to add custom tags to the spans created on the client
side of my thrift calls. Currently this is not possible since the tags
are added in a static method that cannot be extended. Add an extensible
class, ClientSpanDecorator, which can be set on a SpanProtocol which is
used to decorate the spans. The default implementation of
ClientSpanDecorator delegates to the static methods of SpanDecorator to
ensure backwards compatibility. The ClientSpanDecorator can be extended
by consumers of the library when creating SpanProtocols to customise how
the spans are tagged.